+++ esp-hal/CHANGELOG.md
@@ -102,6 +102,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ClockSource` enums are now `#[non_exhaustive]` (#2912)
 
+- `gpio::{Input, Flex}::wakeup_enable` now returns an error instead of panicking. (#2916)
+
 ### Fixed
 
 - Xtensa devices now correctly enable the `esp-hal-procmacros/rtc-slow` feature (#2594)